+++ clang/docs/ReleaseNotes.rst
@@ -404,6 +404,10 @@ Improvements to Clang's diagnostics
 - Clang now emits a diagnostic in case `vector_size` or `ext_vector_type`
   attributes are used with a negative size (#GH165463).
 
+- A new warning, -Wshadow-header, has been added to detect when a header file
+  included via quotes (#include "...") is found in multiple distinct search
+  directories.
+
 Improvements to Clang's time-trace
 ----------------------------------